MIDIBox/requirements.txt
bidict=0.18.2
RPi.GPIO
MFRC522
mido